+++ official/vision/beta/projects/yolo/modeling/building_blocks/dark_tiny.py
@@ -1,7 +1,7 @@
 """Contains common building blocks for yolo neural networks."""
 import tensorflow as tf
 import tensorflow.keras as ks
-from ._DarkConv import DarkConv
+from .dark_conv import DarkConv
 
 
 @ks.utils.register_keras_serializable(package='yolo')
