--- official/vision/beta/projects/yolo/modeling/tests/README.md
@@ -1,1 +1,0 @@
-# Unit Tests for Yolo and Darknet models
